refactor package of aws bulk loader provider
